Backend/requirements.txt
# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.0

# --- Config & Validation ---
pydantic[dotenv]==2.8.0
python-dotenv==1.0.1

# --- File Upload & Parsing ---
python-multipart==0.0.9
pdfplumber==0.11.4
pymupdf==1.24.4
camelot-py==0.11.0
pdfminer.six==20231228

# --- OCR & Image Processing ---
pytesseract==0.3.10
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# --- Database ---
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9   # For PostgreSQL

# --- Vector DB + LLM Integration ---
langchain==0.2.10
langchain-community==0.2.9
langchain-openai==0.1.7
chromadb==0.5.3
sentence-transformers==3.0.1
huggingface_hub==0.24.6

# --- Logging & Monitoring ---
structlog==24.2.0
sentry-sdk==2.8.0

# --- Caching & Task Queue (optional but production-grade) ---
redis==5.0.7
celery==5.4.0

# --- Cloud/File Handling (optional future-proofing) ---
boto3==1.34.139   # AWS S3 client
python-magic==0.4.27  # File type detection

# --- Testing ---
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0   # For API testing

# --- Code Quality ---
black==24.8.0
flake8==7.1.1
mypy==1.11.1
isort==5.13.2

# --- Pre-commit hooks ---
pre-commit==3.8.0
